Added "glyph" hidden layer to the (code-switch) model.
In the new generative story, the language model generates a sequence of
characters, and each LM character generates a "glyph" character, which
represents the character that will be rendered.  This is useful since
some printers used spellings (or character elisions) that differ from
the "true" form of the language that is produced by the LM.  The
characters rendered on the page are thus reflections of the glyph hidden
states, and not the LM states directly.