Commenting out assertion that fails.
The assertion has been in there since the beginning, but was coded in
such a way that no error was thrown when it got tripped.  Once the
assertion mechanism was fixed, errors started getting thrown.  But the
code still works (just as it always has) even those the condition in the
assertion is being violated.  Not sure whether this check is actually
correct and there are bugs elsewhere, or if this is checking a state
that should be considered valid.